update redux, redux-observable, rxjs versions, add function to resolve breaking changes with redux-observable
