+++ .node-version
@@ -1,0 +1,1 @@
+10.13.0
