+++ README.md
@@ -6,33 +6,35 @@
 
 Decorators for Redux Observable
 
-When using Redux with Angular with ng-redux and redux-observable, it's common to create your epics as an injectable class, and when configuring the store - creating an epic middleware for each one, or using combineEpics:
+When using Redux with Angular with __angular-redux/store__ and __redux-observable__, it's common to create your epics as an injectable class, and when configuring the store - creating an epic middleware for each one, or using `combineEpics`:
 
 ```ts
 @Injectable()
 export class SomeEpics {
-	epicOne = (action$) => action$.ofType('PING').mapTo({type: 'PONG'});
-	epicTwo = (action$) => action$.ofType('ACTION_IN').mapTo({type: 'ACTION_OUT'});
+	epicOne = (action$) => action$.ofType('PING').pipe(mapTo({type: 'PONG'}));
+	epicTwo = (action$) => action$.ofType('ACTION_IN').pipe(mapTo({type: 'ACTION_OUT'}));
 }
 
 @NgModule({
 
 })
 export class AppModule {
-	constructor(ngRedux:NgRedux, someEpics:SomeEpics) {
+	constructor(ngRedux: NgRedux, someEpics: SomeEpics) {
 		let epics = combineEpics(
 			someEpics.epicOne,
 			someEpics.epicTwo
-		)
-		
-		ngRedux.configureStore(reducer,[createEpicMidleware(epics)])
-		
-		// or 
+		);
+		let epicMiddleware = createEpicMidleware();
+
+		ngRedux.configureStore(reducer,[epicMiddleware]);
+		epicMiddleware.run(epics);
+
+		// or
 
 		let epicOne = createMiddleware(someEpics.epicOne);
-		let epicTwo = createMiddleware(someEpics.epicOne);
+		let epicTwo = createMiddleware(someEpics.epicTwo);
 
-		ngRedux.configureStore(reducer,[epicOne, epicTwo)])
+		ngRedux.configureStore(reducer,[epicOne, epicTwo]);
 	}
 }
 ```
@@ -40,25 +42,29 @@ export class AppModule {
 This decorator is intended to make it easier to mark which properties / methods in a class are Epics to simplify creating the epic middleware for your application.
 
 ```ts
-import { Epic } from 'redux-observable-decorator'
+import { Epic } from 'redux-observable-decorator';
+
 @Injectable()
 export class SomeEpics {
-	@Epic() epicOne = (action$) => action$.ofType('PING').mapTo({type: 'PONG'});
-	@Epic() epicTwo = (action$) => action$.ofType('ACTION_IN').mapTo({type: 'ACTION_OUT'});
+	@Epic() epicOne = (action$) => action$.ofType('PING').pipe(mapTo({type: 'PONG'}));
+	@Epic() epicTwo = (action$) => action$.ofType('ACTION_IN').pipe(mapTo({type: 'ACTION_OUT'}));
 }
 ```
 
 ```ts
-import { createEpics } from 'redux-observable-decorator';
+import { combineDecoratedEpics } from 'redux-observable-decorator';
+import { createEpicMiddleware } from 'redux-observable';
+
 @NgModule({
 
 })
 export class AppModule {
 	constructor(ngRedux:NgRedux, someEpics:SomeEpics) {
-		let epics = createEpics(someEpics)
-		
-		ngRedux.confgureStore(reducer,[epics])
-	
+		let epics = combineDecoratedEpics(someEpics);
+		const epicMiddleware = createEpicMiddleware();
+
+		ngRedux.configureStore(reducer,[epicMiddleware]);
+		epicMiddleware.run(epics);
 	}
 }
 ```
@@ -67,25 +73,26 @@ This can be used with vanilla redux also - as seen in the unit tests...
 
 ```ts
 class Test {
-	@Epic() epic = (action$) => action$.ofType('TEST_IN').mapTo({ type: 'TEST_OUT' });
+	@Epic() epic = (action$) => action$.ofType('TEST_IN').pipe(mapTo({ type: 'TEST_OUT' }));
 }
 
 const reducer = (state = [], action) => state.concat(action);
 const epics = new Test();
-const epicMiddleware = createEpics(epics);
+const epicMiddleware = createEpicMiddleware(epics);
 const store = createStore(reducer, applyMiddleware(epicMiddleware));
+epicMiddleware.run(combineDecoratedEpics(epics));
 ```
 
 # Inspiration
 
 The `@Effect` decorator from [ngrx/effects](https://github.com/ngrx/effects)
 
-# Todo 
+# Todo
 
 * [ ] Better docs
 * [ ] Publish on NPM
 * [ ] Improve tests
-* [ ] Get test coverage working 
-* [ ] Some Anglar 2 / integration tests 
+* [ ] Get test coverage working
+* [ ] Some Anglar 2 / integration tests
 * [ ] Example App
 * [ ] Strategy for lazy loading epics (to support code-splitting)?
