+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "redux-observable-decorator",
-  "version": "1.2.0-0",
+  "version": "2.0.0",
   "description": "Decorators for Redux Observable",
   "main": "./release/index.js",
   "files": [
@@ -9,12 +9,11 @@
   "typings": "./release/index.d.ts",
   "scripts": {
     "test": "karma start --single-run",
+    "prepare": "npm run build",
     "build": "npm run build:js",
-    "build:js": "ngc -p tsconfig.dist.json",
+    "build:js": "tsc --build tsconfig.dist.json",
     "prebuild": "npm run test && npm run clean:pre",
-    "postbuild": "npm run clean:post",
     "clean:pre": "rimraf release",
-    "clean:post": "rimraf **/*.ngsummary.json",
     "postversion": "npm run build"
   },
   "repository": "git+https://github.com/angular-redux/redux-observable-decorator.git",
@@ -25,42 +24,33 @@
   },
   "homepage": "https://github.com/angular-redux/redux-observable-decorator#readme",
   "peerDependencies": {
-    "redux": "3.*",
-    "redux-observable": "^0.13.0 || ^0.14.0 || ^0.15.0 || ^0.16.0"
+    "redux-observable": "^1.0.0"
   },
   "devDependencies": {
-    "@angular/common": "4.0.0",
-    "@angular/compiler": "4.0.0",
-    "@angular/compiler-cli": "4.0.0",
-    "@angular/core": "4.0.0",
-    "@angular/platform-browser": "4.0.0",
-    "@angular/platform-browser-dynamic": "4.0.0",
-    "@angular/platform-server": "4.0.0",
+    "@types/core-js": "^2.5.0",
     "@types/jasmine": "^2.5.38",
     "@types/node": "^6.0.52",
-    "awesome-typescript-loader": "^3.0.0-beta.17",
-    "core-js": "^2.4.1",
+    "@types/webpack-env": "^1.13.6",
+    "core-js": "^2.6.1",
     "istanbul-instrumenter-loader": "^1.1.0",
     "jasmine-core": "^2.5.2",
-    "karma": "^1.4.1",
-    "karma-chrome-launcher": "^2.0.0",
-    "karma-coverage": "^1.1.1",
-    "karma-jasmine": "^1.1.0",
-    "karma-mocha-reporter": "^2.2.1",
+    "karma": "^3.1.4",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-coverage": "^1.1.2",
+    "karma-jasmine": "^2.0.1",
+    "karma-mocha-reporter": "^2.2.5",
     "karma-sourcemap-loader": "^0.3.7",
-    "karma-typescript-preprocessor": "^0.3.1",
-    "karma-webpack": "^2.0.2",
-    "redux": "^3.6.0",
-    "redux-observable": "^0.16.0",
+    "karma-webpack": "^4.0.0-rc.5",
+    "redux": "^4.0.1",
+    "redux-observable": "^1.0.0",
     "reflect-metadata": "^0.1.8",
     "rimraf": "^2.5.4",
-    "rxjs": "^5.0.1",
-    "ts-loader": "^2.2.7",
+    "rxjs": "^6.3.3",
+    "ts-loader": "^5.3.2",
     "tslint": "^4.1.1",
-    "tslint-loader": "^3.3.0",
-    "typescript": "^2.4.0",
-    "uglifyjs": "^2.4.10",
-    "webpack": "^2.2.1",
-    "zone.js": "^0.8.4"
+    "tslint-loader": "^3.5.4",
+    "typescript": "^3.2.2",
+    "webpack": "^4.28.3",
+    "zone.js": "^0.8.26"
   }
 }
