+++ tsconfig.dist.json
@@ -1,27 +1,16 @@
 {
+  "extends": "./tsconfig.json",
   "compilerOptions": {
-    "baseUrl": ".",
+    "sourceMap": true,
     "declaration": true,
     "stripInternal": true,
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
-    "module": "commonjs",
-    "moduleResolution": "node",
     "noEmitOnError": false,
-    "outDir": "./release",
     "rootDir": "./src",
-    "sourceMap": true,
     "inlineSources": true,
-    "lib": ["es2015", "dom"],
-    "target": "es5",
     "skipLibCheck": true
   },
   "files": [
     "src/index.ts"
-  ],
-  "exclude": [
-    "node_modules",
-    "spec"
   ],
   "angularCompilerOptions": {
     "strictMetadataEmit": true
