+++ tsconfig.json
@@ -1,27 +1,19 @@
 {
-  "version": "2.0.0",
   "compileOnSave": false,
   "compilerOptions": {
+    "baseUrl": ".",
     "outDir": "./release",
     "target": "es5",
     "sourceMap": false,
     "emitDecoratorMetadata": true,
     "experimentalDecorators": true,
+    "module": "commonjs",
     "removeComments": false,
     "moduleResolution": "node",
-    "declaration": true,
-    "module": "commonjs",
-    "lib": ["es5", "dom", "es2015"],
-    "types": ["jasmine", "node"],
-    "baseUrl": ".",
+    "lib": ["dom","es2015"],
     "rootDir": ".",
     "paths": {
       "@redux-observable-decorator": ["./"]
     }
-  },
-  "exclude": ["node_modules", "release", "tests.bundle.ts"],
-  "angularCompilerOptions": {
-    "genDir": ".ngc"
-  },
-  "types": ["jasmine", "node"]
+  }
 }
